Completed search cities screen main functionality

+++ core/src/main/java/cat/ppicas/cleanarch/domain/CurrentWeather.java
@@ -3,16 +3,16 @@ package cat.ppicas.cleanarch.domain;
 public class CurrentWeather {
 
     private String mCityId;
-    private int mCurrentTemp;
-    private Integer mMaxTemp;
-    private Integer mMinTemp;
+    private double mCurrentTemp;
+    private Double mMaxTemp;
+    private Double mMinTemp;
 
-    public CurrentWeather(String cityId, int currentTemp) {
+    public CurrentWeather(String cityId, double currentTemp) {
         mCityId = cityId;
         mCurrentTemp = currentTemp;
     }
 
-    public CurrentWeather(String cityId, int currentTemp, int maxTemp, int minTemp) {
+    public CurrentWeather(String cityId, double currentTemp, double maxTemp, double minTemp) {
         mCityId = cityId;
         mCurrentTemp = currentTemp;
         mMaxTemp = maxTemp;
@@ -23,15 +23,15 @@ public class CurrentWeather {
         return mCityId;
     }
 
-    public int getCurrentTemp() {
+    public double getCurrentTemp() {
         return mCurrentTemp;
     }
 
-    public Integer getMaxTemp() {
+    public Double getMaxTemp() {
         return mMaxTemp;
     }
 
-    public Integer getMinTemp() {
+    public Double getMinTemp() {
         return mMinTemp;
     }
 }
